core/analysis/src/main/java/com/buschmais/jqassistant/core/analysis/impl/RowCountVerificationStrategy.java
package com.buschmais.jqassistant.core.analysis.impl;

import java.util.List;
import java.util.Map;

import com.buschmais.jqassistant.core.analysis.api.Result;
import com.buschmais.jqassistant.core.rule.api.model.ExecutableRule;
import com.buschmais.jqassistant.core.rule.api.reader.RowCountVerification;

public class RowCountVerificationStrategy extends AbstractMinMaxVerificationStrategy implements VerificationStrategy<RowCountVerification> {

    @Override
    public Class<RowCountVerification> getVerificationType() {
        return RowCountVerification.class;
    }

    @Override
    public <T extends ExecutableRule> Result.Status verify(T executable, RowCountVerification verification, List<String> columnNames,
            List<Map<String, Object>> rows) {
        return getStatus(executable, rows.size(), verification.getMin(), verification.getMax());
    }

}
